Added guest user for current user
