+++ app/controllers/application_controller.rb
@@ -2,13 +2,5 @@ class ApplicationController < ActionController::Base
   protect_from_forgery with: :exception
   include DeviseWhitelist
   include SetSource
-
-  def current_user
-    super || OpenStruct.new( 
-      name: "Guest User", 
-      first_name: "Guest", 
-      last_name: "User", 
-      email: "guest@example.com" 
-      )
-  end
+  include CurrentUserConcern
 end
