+++ app/controllers/concerns/current_user_concern.rb
@@ -1,0 +1,16 @@
+module CurrentUserConcern
+  extend ActiveSupport::Concern
+  
+  def current_user
+      super || guest_user
+  end
+
+  def guest_user
+    OpenStruct.new(
+      name: "Guest User",
+      first_name: "Guest", 
+      last_name: "User", 
+      email: "guest@example.com"
+      )
+  end
+end
